requirements.txt
argparse==1.2.1
backports-abc==0.5
backports.ssl-match-hostname==3.5.0.1
certifi==2016.9.26
cffi==1.9.1
cryptography==1.7.1
docutils==0.13.1
enum34==1.1.6
idna==2.2
ipaddress==1.0.18
lockfile==0.12.2
luigi==2.5.0
nose==1.3.7
pyOpenSSL==16.2.0
pyasn1==0.1.9
pycparser==2.17
python-daemon==2.1.2
requests==2.12.5
requirements-parser==0.1.0
simplejson==3.10.0
singledispatch==3.4.0.3
six==1.10.0
tornado==4.4.2
urllib3==1.20
wsgiref==0.1.2
